+++ Wiki/Azure/CHANGELOG.md
@@ -1,8 +1,17 @@
 [[_TOC_]]
 
 # Change Log
-
 All notable changes to this project will be documented in this file.
+
+## [2022.07.13]
+- Added script `Create-CDN-Custom-Domain`
+- Added script `Create-CDN-Endpoint-Rule`
+- Added script `Create-CDN-Endpoint`
+- Added script `Create-CDN-Profile`
+
+## [2022.06.20]
+- Added script `Get-CosmosDb-Connectionstring.ps1`
+- Updated script `Create-CosmosDbAccount.ps1` to add the capabilities parameter as an optional parameter during creation.
 
 ## [2022.06.14]